+++ lib/graphite_async.mli
@@ -27,6 +27,7 @@ val set : t -> key:string -> int -> unit
 val remove : t -> key:string -> unit
 val add_percentile_observation : t -> key:string -> Percentile.value -> unit
 val add_percentile_observation_opt : t option -> key:string -> Percentile.value -> unit
+val get_report : t -> string
 
 module Result : sig
   val time :